refactor: extract bootstrap from App
