+++ src/app/App.tsx
@@ -224,6 +224,8 @@ import { createNoteActions } from "./note-actions";
 
 import { useDictationEvents } from "./use-dictation-events";
 
+import { useAppBootstrap } from "./use-app-bootstrap";
+
 export function App() {
   const replayOnboarding = shouldReplayOnboarding();
   const activeHermesProfileName = useActiveHermesProfileName();
@@ -1548,55 +1550,18 @@ export function App() {
     void dictationHelperCommand({ type: "get_permission_status" }).catch(() => undefined);
   }, []);
 
-  useEffect(() => {
-    if (appBlocked) return;
-    bootstrapApp()
-      .then(async (payload) => {
-        const seeded = withFakeRecovery(payload);
-        dispatch({ type: "bootstrapLoaded", payload: seeded.payload });
-        const activeRecording = seeded.payload.activeRecording;
-        const activeRecordingNoteId = activeRecording?.noteId;
-        if (activeRecording) {
-          recordingStatusRef.current = activeRecording;
-          dispatch({ type: "recordingStatusChanged", status: activeRecording });
-          if (activeRecordingNoteId) {
-            calendarContextNoteProfilesRef.current.set(
-              activeRecordingNoteId,
-              getActiveHermesProfileName(),
-            );
-            pendingCalendarContextAdoptionsRef.current.add(activeRecordingNoteId);
-            setRecordingNote(activeRecordingNoteId);
-          }
-        }
-        if (seeded.fakeNote) {
-          dispatch({ type: "noteLoaded", note: seeded.fakeNote });
-          // The fake-recovery dev flow inspects the notes list, so it skips
-          // the agent landing.
-          setActiveView("notes");
-          setBootstrapped(true);
-          return;
-        }
-        // The app lands on the agent view, but a note is still selected up
-        // front. After a webview reload, a live native capture takes priority
-        // so its recorder controls and note association are restored.
-        if (seeded.payload.notes.length === 0 && !activeRecordingNoteId) {
-          const note = await createNote(undefined);
-          dispatch({ type: "noteLoaded", note });
-          setBootstrapped(true);
-          return;
-        }
-        const firstNoteId = activeRecordingNoteId ?? seeded.payload.notes[0]?.id;
-        if (firstNoteId) {
-          const note = await getNote(firstNoteId);
-          const calendarUpdate = calendarContextNoteUpdatesRef.current.get(firstNoteId);
-          calendarContextNoteUpdatesRef.current.delete(firstNoteId);
-          pendingCalendarContextAdoptionsRef.current.delete(firstNoteId);
-          dispatch({ type: "noteLoaded", note: calendarUpdate ?? note });
-        }
-        setBootstrapped(true);
-      })
-      .catch((err: unknown) => setError(messageFromError(err)));
-  }, [appBlocked]);
+  useAppBootstrap({
+    appBlocked,
+    calendarContextNoteProfilesRef,
+    calendarContextNoteUpdatesRef,
+    dispatch,
+    pendingCalendarContextAdoptionsRef,
+    recordingStatusRef,
+    setActiveView,
+    setBootstrapped,
+    setError,
+    setRecordingNote,
+  });
 
   useEffect(() => {
     function handleProfileDataChanged(event: Event) {
